fix: move controllers back to mandatory location (#885)
