+++ src/Model/Settings/AccountSettings.php
@@ -7,11 +7,7 @@ namespace MyParcelNL\Magento\Model\Settings;
 use Exception;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\ObjectManager;
-use MyParcelNL\Magento\Controller\Adminhtml\Settings\CarrierConfigurationImport;
 use MyParcelNL\Magento\Service\Config;
-use MyParcelNL\Sdk\Exception\AccountNotActiveException;
-use MyParcelNL\Sdk\Exception\ApiException;
-use MyParcelNL\Sdk\Exception\MissingFieldException;
 use MyParcelNL\Sdk\Factory\Account\CarrierConfigurationFactory;
 use MyParcelNL\Sdk\Model\Account\Account;
 use MyParcelNL\Sdk\Model\Account\CarrierConfiguration;
